requirements.txt
asgiref==3.2.10
Django==3.1.2
django-cors-headers==3.5.0
django-filter==2.4.0
djangorestframework==3.12.1
importlib-metadata==2.0.0
Markdown==3.3.3
Pillow==8.0.1
pytz==2020.1
sqlparse==0.4.1
zipp==3.4.0
